refactor: move the utils file to the stats folder
